# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API VSM Mobile",
        "description": "Documentation API VSM PK11",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur Local"
        }
    ],
    "paths": {
...
        "/contributions/my-status": {
            "get": {
                "tags": [
                    "Trésorerie"
                ],
                "summary": "Statut des cotisations du membre connecté",
                "operationId": "6f752b66731d970728f9874242b729f8",
                "responses": {
                    "200": {
                        "description": "Historique et statut individuel du joueur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "monthly_fee": {
                                                    "type": "integer",
                                                    "example": 10000
                                                },
                                                "is_up_to_date": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "last_payment_date": {
                                                    "type": "string",
                                                    "example": "2026-08-01",
                                                    "nullable": true
                                                },
                                                "balance_due": {
                                                    "type": "integer",
                                                    "example": 0
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/contributions": {
            "get": {
                "tags": [
                    "Trésorerie"
                ],
                "summary": "Liste globale des cotisations (Trésorier/Admin)",
                "operationId": "f7ef2ad60f4efed1610da7a2f1275d48",
                "parameters": [
                    {
                        "name": "month",
                        "in": "query",
                        "description": "Filtrer par mois (ex: 2026-08)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "2026-08"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filtrer par membre spécifique",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bilan global des cotisations et transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_amount": {
                                                    "type": "integer",
                                                    "example": 450000
                                                },
                                                "records": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "user_name": {
                                                                "type": "string",
                                                                "example": "Emmanuel Dika"
                                                            },
                                                            "amount": {
                                                                "type": "integer",
                                                                "example": 10000
                                                            },
                                                            "payment_type": {
                                                                "type": "string",
                                                                "example": "monthly_fee"
                                                            },
                                                            "payment_method": {
                                                                "type": "string",
                                                                "example": "orange_money"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "example": "2026-08-01 10:30:00"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Trésorerie"
                ],
                "summary": "Enregistrer un paiement de cotisation (Trésorier/Admin)",
                "operationId": "82d74ba2480dbc4035a8ed440c7accb5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "amount",
                                    "payment_type",
                                    "payment_method"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "example": 10000
                                    },
                                    "payment_type": {
                                        "type": "string",
                                        "example": "monthly_fee",
                                        "enum": [
                                            "monthly_fee",
                                            "event_fee",
                                            "donation"
                                        ]
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "orange_money",
                                        "enum": [
                                            "cash",
                                            "orange_money",
                                            "mtn_momo"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Cotisation du mois d'Août 2026"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paiement enregistré avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation des données"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/contributions/{id}": {
            "put": {
                "tags": [
                    "Trésorerie"
                ],
                "summary": "Mettre à jour un enregistrement de cotisation",
                "operationId": "b6f320a094c43d9be3bd64b2c73c36e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la transaction",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "integer",
                                        "example": 15000
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "cash"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Ajustement du montant"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction mise à jour"
                    },
                    "404": {
                        "description": "Transaction introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Entrez votre jeton Sanctum sous la forme: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}